requirements.txt
alembic
APScheduler
asgiref
bcrypt
bidict
blinker
boolean.py
boto3
botocore
CacheControl
cachelib
certifi
cffi
charset-normalizer
click
colorama
cryptography
cyclonedx-python-lib
defusedxml
Deprecated
distutils-pytest
Django
dnspython
docopt
dsutils
email_validator
et_xmlfile
eventlet
filelock
Flask
Flask-Caching
Flask-Dance
Flask-Limiter
Flask-Login
Flask-Migrate
Flask-SocketIO
Flask-SQLAlchemy
Flask-WTF
greenlet
h11
idna
iniconfig
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
Jinja2
jmespath
joblib
keyring
license-expression
limits
Mako
Markdown
markdown-it-py
MarkupSafe
mdurl
more-itertools
msgpack
netaddr
npm
numpy
oauthlib
openpyxl
optional-django
ordered-set
packageurl-python
packaging
pandas
passlib
pendulum
pip-api
pip-requirements-parser
pip_audit
pipreqs
platformdirs
pluggy
psycopg2
psycopg2-binary
py-serializable
pycountry
pycparser
Pygments
PyJWT
pyparsing
pytest
python-dateutil
python-dotenv
python-engineio
python-socketio
python3-openid
pytz
pywin32-ctypes
requests
requests-oauthlib
rich
s3transfer
schedule
scikit-learn
scipy
setuptools
simple-websocket
six
social
social-auth-app-django
social-auth-core
sortedcontainers
SQLAlchemy
sqlparse
threadpoolctl
toml
typing_extensions
tzdata
tzlocal
urllib3
URLObject
Werkzeug
wrapt
wsproto
WTForms
yarg